refactor(vscode-core): own the storage shim registration once
SonarQube reported two more blocks of the same shape commit 9f869ba4b
already fixed for phase 2: container.ts against phase-1-infra.ts (24
lines) and against phase-3-storage.ts (17 lines).

Both register vscode-core-owned tokens over platform-core ports, so the
registration fact moved into vscode-core as registerExtensionContextShim
and registerStateStorageAdapters. Both composition roots now call them.
Net 108 lines removed across the two roots.

No new import edge. Both ptah-electron and cli-engine already import
vscode-core, and ptah-electron does not reach into cli-engine. The full
CLI and TUI esbuild bundle builds unchanged, so neither tsconfig.build.json
paths allowlist needed an entry.

The VS Code host has no copy of either block, since it receives a real
vscode.ExtensionContext. The helper says so.

Two cosmetic behaviour deltas: the log prefix is now [VS Code Core], the
precedent set by registerTypeScriptDiagnosticsProvider, and the Electron
error path logs the Error object rather than its message alone, which is
the CLI's richer form.

Boot order is unchanged. Nothing moved between phases and nothing changed
position within one. boot-order, container.smoke, phase-2-diagnostics-override
and rpc-surface all pass unmodified, and expected-resolvable.ts needed no
edit because it lists only RPC handler classes.

The 22-line residual between the two trigger services is left alone. Half
of it is the call site of the already-extracted BootScanScheduler, and the
other half diverges on the line after the block ends. Sharing it would mean
three new structural ports and a duplicate options type to save fourteen
lines.
