+++ libs/backend/vscode-core/src/index.ts
@@ -2,6 +2,11 @@ export { TOKENS } from './di/tokens';
 export { registerVsCodeCoreServices } from './di';
 export { registerVsCodeCorePlatformAgnostic } from './di/register-platform-agnostic';
 export type { PlatformAgnosticRegistrationOptions } from './di/register-platform-agnostic';
+export {
+  registerExtensionContextShim,
+  registerStateStorageAdapters,
+} from './di/register-storage-shims';
+export type { ExtensionContextShimPaths } from './di/register-storage-shims';
 export { Logger } from './logging';
 export type { LogLevel, LogContext, LogEntry } from './logging';
